address Sam's feedback
