+++ lib/packet/spse/scmp_auth/ext_hashtree.py
@@ -1,0 +1,167 @@
+# Copyright 2017 ETH Zurich
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+:mod:`extn` --- SCMPAuthDRKey and SCMPAuthHashTree extension header
+=================================================================
+"""
+# Stdlib
+import struct
+
+# SCION
+from lib.packet.spse.defines import (
+    SPSEValidationError,
+    SPSELengths,
+    SPSESecModes,
+)
+from lib.packet.spse.ext import SCIONPacketSecurityBaseExtn
+from lib.types import ExtEndToEndType
+from lib.util import hex_str, Raw
+
+
+class SCMPAuthHashTreeExtn(SCIONPacketSecurityBaseExtn):
+    """
+    Implementation of the SCMPAuthHashTree extension, which is based on the
+    SCIONPacketSecurity extension.
+
+    0B       1        2        3        4        5        6        7
+    +--------+--------+--------+--------+--------+--------+--------+--------+
+    | xxxxxxxxxxxxxxxxxxxxxxxx |  0x05  | Height |            Order         |
+    +--------+--------+--------+--------+--------+--------+--------+--------+
+    |                               Signature (8 lines)                     |
+    +--------+--------+--------+--------+--------+--------+--------+--------+
+    |                               Hashes (height * 2)                     |
+    +--------+--------+--------+--------+--------+--------+--------+--------+
+
+    Height: Height of the hash tree. MAX_HEIGHT is 24.
+    Order: Bit vector. The bit at index i is associated with hash i.
+        0 (1) indicates hash i shall be used as left (right) input.
+    Signature: Signature of resulting hash
+    Hashes: Hashes used to verify the proof. At index 0 is the leaf hash, at
+        index height is the root hash.
+    """
+    NAME = "SCMPAuthHashTreeExtn"
+    EXT_TYPE = ExtEndToEndType.SPSE
+
+    def __init__(self, raw=None):  # pragma: no cover
+        """
+        :param bytes raw: Raw data holding height, order, signature and hashes.
+        """
+        self.height = 0
+        self.order = []
+        self.signature = []
+        self.hashes = []
+        super().__init__(raw)
+        self.sec_mode = SPSESecModes.SCMP_AUTH_HASH_TREE
+
+    def _parse(self, raw):
+        """
+        Parse payload to extract values.
+
+        :param bytes raw: raw payload.
+        """
+        data = Raw(raw, self.NAME)
+        super()._parse(data)
+
+        self.sec_mode = data.pop(SPSELengths.SECMODE)
+        self.height = data.pop(SCMPAuthHashTreeExtn.Lengths.HEIGHT)
+        self.order = data.pop(SCMPAuthHashTreeExtn.Lengths.ORDER)
+        self.signature = data.pop(SCMPAuthHashTreeExtn.Lengths.SIGNATURE)
+        self.hashes = data.pop(self.height * SCMPAuthHashTreeExtn.Lengths.HASH)
+
+    @classmethod
+    def from_values(cls, height, order, signature, hashes):  # pragma: no cover
+        """
+        Construct extension.
+
+        :param int height: Height of the hash tree.
+        :param bytes order: bit vector indicating left or right hash.
+        :param bytes signature: Signature of the resulting hash.
+        :param bytes hashes: Hashes needed to conduct the proof.
+        :returns: The created instance.
+        :rtype: SCMPAuthHashTreeExtn
+        :raises: SPSEValidationError
+        """
+        error = cls.check_validity(height, order, signature, hashes)
+        if error:
+            raise error
+
+        inst = cls()
+        inst.height = height
+        inst.order = order
+        inst.signature = signature
+        inst.hashes = hashes
+        hdr_len = inst.bytes_to_hdr_len(
+            SCMPAuthHashTreeExtn.Lengths.HASH_TREE_MIN_LENGTH +
+            height * SCMPAuthHashTreeExtn.Lengths.HASH)
+        inst._init_size(hdr_len)
+        return inst
+
+    def pack(self):
+        """
+        Pack extension into byte string
+
+        :returns: packed extension.
+        :rtype: bytes
+        """
+        packed = [struct.pack("!B", self.sec_mode),
+                  struct.pack("!B", self.height),
+                  self.order,
+                  self.signature,
+                  self.hashes]
+        raw = b"".join(packed)
+        self._check_len(raw)
+        return raw
+
+    @staticmethod
+    def check_validity(height, order, signature, hashes):
+        """
+        Check if parameters are valid
+
+        :param int height: Height of the hash tree.
+        :param bytes order: bit vector indicating left or right hash.
+        :param bytes signature: Signature of the resulting hash.
+        :param bytes hashes: Hashes needed to conduct the proof.
+        :raises: SPSEValidationError
+        """
+
+        if not 0 <= height <= SCMPAuthHashTreeExtn.MAX_HEIGHT:
+            raise SPSEValidationError("Invalid height %s. Max height %s" % (
+                height, SCMPAuthHashTreeExtn.MAX_HEIGHT))
+        if len(order) != SCMPAuthHashTreeExtn.Lengths.ORDER:
+            raise SPSEValidationError("Invalid order length %s. Expected %s" % (
+                len(order), SCMPAuthHashTreeExtn.Lengths.ORDER))
+        if len(signature) != SCMPAuthHashTreeExtn.Lengths.SIGNATURE:
+            raise SPSEValidationError("Invalid signature length %s. Expected %s" % (
+                len(signature), SCMPAuthHashTreeExtn.Lengths.SIGNATURE))
+        if len(hashes) != height * SCMPAuthHashTreeExtn.Lengths.HASH:
+            raise SPSEValidationError("Invalid order length %s. Expected %s" % (
+                len(hashes), height * SCMPAuthHashTreeExtn.Lengths.HASH))
+
+    def __str__(self):
+        return ("%s(%sB): Height: %s Order: %s\n\tSignature: %s\n\tHashes: %s" % (
+                 self.NAME, len(self), self.height, hex_str(self.order),
+                 hex_str(self.signature), hex_str(self.hashes)))
+
+    # max height of the hash tree
+    MAX_HEIGHT = 24
+
+    class Lengths:
+        """
+        Constant lengths.
+        """
+        HASH = 16
+        HEIGHT = 1
+        ORDER = 3
+        SIGNATURE = 64
+        HASH_TREE_MIN_LENGTH = SPSELengths.SECMODE + HEIGHT + ORDER + SIGNATURE
